Use the `smoke` build tag for smoke tests.
