+++ src/github.com/mozilla-services/pushgo/simplepush/api_smoke_test.go
@@ -1,3 +1,5 @@
+// +build smoke
+
 /* This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
