+++ src/github.com/mozilla-services/pushgo/simplepush/gomemc_store_smoke_test.go
@@ -1,3 +1,4 @@
+// +build smoke
 // +build memcached_server_test
 // +build !cgo !libmemcached
