+++ src/github.com/mozilla-services/pushgo/simplepush/handlers_test.go
@@ -12,11 +12,8 @@ import (
 	"net/http/httptest"
 	"net/url"
 	"strconv"
-	"strings"
 	"testing"
 	"time"
-
-	"github.com/mozilla-services/pushgo/client"
 )
 
 func newTestHandler(tb TBLoggingInterface) *Application {
@@ -187,99 +184,4 @@ func Test_UpdateHandler(t *testing.T) {
 		t.Error("Returned version does not match expected value")
 	}
 
-}
-
-func endpointIds(uri *url.URL) (deviceId, channelId string, ok bool) {
-	if !uri.IsAbs() {
-		ok = false
-		return
-	}
-	pathPrefix := "/update/"
-	i := strings.Index(uri.Path, pathPrefix)
-	if i < 0 {
-		ok = false
-		return
-	}
-	key := strings.SplitN(uri.Path[i+len(pathPrefix):], ".", 2)
-	if len(key) < 2 {
-		ok = false
-		return
-	}
-	return key[0], key[1], true
-}
-
-func TestBadKey(t *testing.T) {
-	origin, err := testServer.Origin()
-	if err != nil {
-		t.Fatalf("Error initializing test server: %#v", err)
-	}
-	conn, deviceId, err := client.Dial(origin)
-	if err != nil {
-		t.Fatalf("Error dialing origin: %#v", err)
-	}
-	defer conn.Close()
-	channelId, endpoint, err := conn.Subscribe()
-	if err != nil {
-		t.Fatalf("Error subscribing to channel: %#v", err)
-	}
-	uri, err := url.ParseRequestURI(endpoint)
-	if err != nil {
-		t.Fatalf("Error parsing push endpoint %#v: %#v", endpoint, err)
-	}
-	keyDevice, keyChannel, ok := endpointIds(uri)
-	if !ok {
-		t.Errorf("Incomplete push endpoint: %#v", endpoint)
-	}
-	if keyDevice != deviceId {
-		t.Errorf("Mismatched device IDs: got %#v; want %#v", keyDevice, deviceId)
-	}
-	if keyChannel != channelId {
-		t.Errorf("Mismatched channel IDs: got %#v; want %#v", keyChannel, channelId)
-	}
-	newURI, _ := uri.Parse(fmt.Sprintf("/update/%s", keyDevice))
-	err = client.Notify(newURI.String(), 1)
-	clientErr, ok := err.(client.Error)
-	if !ok {
-		t.Errorf("Type assertion failed for endpoint error: %#v", err)
-	} else if clientErr.Status() != 404 {
-		t.Errorf("Unexpected endpoint status: got %#v; want 404", clientErr.Status())
-	}
-}
-
-func TestMissingKey(t *testing.T) {
-	origin, err := testServer.Origin()
-	if err != nil {
-		t.Fatalf("Error initializing test server: %#v", err)
-	}
-	conn, deviceId, err := client.Dial(origin)
-	if err != nil {
-		t.Fatalf("Error dialing origin: %#v", err)
-	}
-	defer conn.Close()
-	channelId, endpoint, err := conn.Subscribe()
-	if err != nil {
-		t.Fatalf("Error subscribing to channel: %#v", err)
-	}
-	uri, err := url.ParseRequestURI(endpoint)
-	if err != nil {
-		t.Fatalf("Error parsing push endpoint %#v: %#v", endpoint, err)
-	}
-	keyDevice, keyChannel, ok := endpointIds(uri)
-	if !ok {
-		t.Errorf("Incomplete push endpoint: %#v", endpoint)
-	}
-	if keyDevice != deviceId {
-		t.Errorf("Mismatched device IDs: got %#v; want %#v", keyDevice, deviceId)
-	}
-	if keyChannel != channelId {
-		t.Errorf("Mismatched channel IDs: got %#v; want %#v", keyChannel, channelId)
-	}
-	newURI, _ := uri.Parse(fmt.Sprintf("/update/%s.", keyDevice))
-	err = client.Notify(newURI.String(), 1)
-	clientErr, ok := err.(client.Error)
-	if !ok {
-		t.Errorf("Type assertion failed for endpoint error: %#v", err)
-	} else if clientErr.Status() != 404 {
-		t.Errorf("Unexpected endpoint status: got %#v; want 404", clientErr.Status())
-	}
 }
