+++ src/github.com/mozilla-services/pushgo/simplepush/no_store_smoke_test.go
@@ -1,3 +1,4 @@
+// +build smoke
 // +build !memcached_server_test
 
 /* This Source Code Form is subject to the terms of the Mozilla Public
